........
  r10137 | martineau | 2008-02-15 10:12:11 -0500 (Fri, 15 Feb 2008) | 5 lines
  
  BugId: 3992
  * server-src/amadmin.c: Use 'long long' and '%lld' to sscanf.
  * file server-src/infofile.c: Use %jd to print intmax_t value.
  * file server-src/driver.c: Use 'long long' and '%lld' to sscanf.
........


git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@798 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/amadmin.c
@@ -1701,7 +1701,7 @@ import_one(void)
     char *hostname = NULL;
     char *diskname = NULL;
     long long off_t_tmp;
-    intmax_t time_t_tmp;
+    long long time_t_tmp;
 
     memset(&info, 0, SIZEOF(info_t));
 
@@ -1831,14 +1831,14 @@ import_one(void)
 	skip_integer(s, ch);
 
 	skip_whitespace(s, ch);
-	if(ch == '\0' || sscanf(s - 1, "%ju", &time_t_tmp) != 1) {
+	if(ch == '\0' || sscanf(s - 1, "%lld", &time_t_tmp) != 1) {
 	    goto parse_err;
 	}
         onestat.secs = (time_t)time_t_tmp;
 	skip_integer(s, ch);
 
 	skip_whitespace(s, ch);
-	if(ch == '\0' || sscanf(s - 1, "%ju", &time_t_tmp) != 1) {
+	if(ch == '\0' || sscanf(s - 1, "%lld", &time_t_tmp) != 1) {
 	    goto parse_err;
 	}
 	/* time_t not guarranteed to be long */
@@ -1908,7 +1908,7 @@ import_one(void)
 	skip_integer(s, ch);
 
 	skip_whitespace(s, ch);
-	if((ch == '\0') || sscanf((s - 1), "%ju", &time_t_tmp) != 1) {
+	if((ch == '\0') || sscanf((s - 1), "%lld", &time_t_tmp) != 1) {
 	    break;
 	}
 	/* time_t not guarranteed to be long */
